Add files via upload
So this is a Wordle just like The New York Times but in Bosnian. Player can choose between light and dark mode and there's a stopwatch as well so the player can know how much time it took him to guess the right word. There are a few problems, the alert message won't show after the player has guessed the word correctly and the stopwatch has stooped. Also the special characters "č,ć,š,đ,ž" can only be typed in with the displayed keyboard and I didn't know how to add "lj,nj,dž" since those characters are made from two letters. But all in all it's good for a small project.